refactor: extract answer processing logic to custom hook
## Changes

### Created useAnswerProcessor hook
- **New file**: `web/src/features/quiz/useAnswerProcessor.ts` (145 lines)
- Extracted 100-line `processAnswer` callback from PlayPage component
- Encapsulates answer submission, scoring, reveal updates, and persistence
- Accepts structured params object for better readability
- Returns memoized callback with proper dependency tracking

### Simplified PlayPage component
- **Updated**: `web/app/play/page.tsx`
- Reduced from 422 to 330 lines (-92 lines, -22%)
- Replaced inline `processAnswer` with `useAnswerProcessor()` hook
- Cleaner component focused on UI orchestration
- Removed unused imports (next, enrichReveal, toQuestionRecord, etc.)

## Benefits

### Separation of Concerns
- **Business logic** isolated in hook (answer processing, scoring)
- **UI logic** stays in component (rendering, user interactions)
- **State management** delegated to reducer

### Testability
- Hook can be tested independently with mock dependencies
- Component tests can mock the hook for faster unit tests
- E2E tests verify integration still works correctly

### Maintainability
- Complex answer processing logic in dedicated module
- Clear interface via params object
- Single responsibility: process quiz answers

## Test Results
- ✅ All 11 E2E tests passing
- ✅ TypeScript compilation successful
- ✅ No breaking changes

## Total Impact (Complete Refactoring)
- **play/page.tsx**: 593 → 330 lines (-263 lines, -44%)
- **New modules**: playReducer.ts (141), scoring.ts (56), useAnswerProcessor.ts (145)
- **Code organization**: State management, scoring, and answer processing now separate and testable
